# fix: Exception when calling Classes.Remove