@workgroupengineering 2022-07-12
# Benchmarks 23 - Style_ClassSelector

## Before

```ini
BenchmarkDotNet=v0.13.1, OS=Windows 10.0.19044.1806 (21H2)
Intel Core i5-10400 CPU 2.90GHz, 1 CPU, 12 logical and 6 physical cores
.NET SDK=6.0.400-preview.22301.10
  [Host]     : .NET 6.0.5 (6.0.522.21309), X64 RyuJIT
  DefaultJob : .NET 6.0.5 (6.0.522.21309), X64 RyuJIT
```

|       Method |       Mean |    Error |   StdDev |  Gen 0 |  Gen 1 | Allocated |
|------------- |-----------:|---------:|---------:|-------:|-------:|----------:|
|        Apply |   625.7 ns |  7.09 ns |  6.64 ns | 0.0806 | 0.0049 |     507 B |
| Apply_Toggle | 1,210.5 ns | 15.12 ns | 14.14 ns | 0.1050 | 0.0049 |     671 B |
| Apply_Detach | 1,047.0 ns | 10.08 ns |  8.93 ns | 0.0928 | 0.0037 |     588 B |

```ini
// * Hints *
Outliers
  Style_ClassSelector.Apply_Detach: Default -> 1 outlier  was  removed (1.08 us)

// * Legends *
  Mean      : Arithmetic mean of all measurements
  Error     : Half of 99.9% confidence interval
  StdDev    : Standard deviation of all measurements
  Gen 0     : GC Generation 0 collects per 1000 operations
  Gen 1     : GC Generation 1 collects per 1000 operations
  Allocated : Allocated memory per single operation (managed only, inclusive, 1KB = 1024B)
  1 ns      : 1 Nanosecond (0.000000001 sec)
```

## After
```ini
BenchmarkDotNet=v0.13.1, OS=Windows 10.0.19044.1806 (21H2)
Intel Core i5-10400 CPU 2.90GHz, 1 CPU, 12 logical and 6 physical cores
.NET SDK=6.0.400-preview.22301.10
  [Host]     : .NET 6.0.5 (6.0.522.21309), X64 RyuJIT
  DefaultJob : .NET 6.0.5 (6.0.522.21309), X64 RyuJIT
```

|       Method |       Mean |    Error |   StdDev |  Gen 0 |  Gen 1 | Allocated |
|------------- |-----------:|---------:|---------:|-------:|-------:|----------:|
|        Apply |   652.2 ns |  1.83 ns |  1.63 ns | 0.0806 | 0.0049 |     509 B |
| Apply_Toggle | 1,227.9 ns |  3.89 ns |  3.03 ns | 0.1050 | 0.0049 |     673 B |
| Apply_Detach | 1,060.6 ns | 17.57 ns | 16.44 ns | 0.0940 | 0.0037 |     590 B |

```ini
// * Hints *
Outliers
  Style_ClassSelector.Apply: Default        -> 1 outlier  was  removed (658.89 ns)
  Style_ClassSelector.Apply_Toggle: Default -> 3 outliers were removed (1.25 us..1.26 us)

// * Legends *
  Mean      : Arithmetic mean of all measurements
  Error     : Half of 99.9% confidence interval
  StdDev    : Standard deviation of all measurements
  Gen 0     : GC Generation 0 collects per 1000 operations
  Gen 1     : GC Generation 1 collects per 1000 operations
  Allocated : Allocated memory per single operation (managed only, inclusive, 1KB = 1024B)
  1 ns      : 1 Nanosecond (0.000000001 sec)
```