@grokys 2022-07-12
Hi @workgroupengineering, sorry but I'm already working on this. We need a generic solution to this problem as it crops up in a few places, with unit tests etc. Also the `ReaderWriterLockSlim` stuff is unnecessary because the listeners should only be added/removed on the UI thread - the problem is that of re-entrancy not thread safety.