@MarchingCube 2022-07-12
For me this is an instant repro in master, I will take a look.