@danwalmsley 2022-07-12
https://github.com/AvaloniaUI/Avalonia/pull/8492