@grokys 2022-07-12
> this issue can simple fix with this code:

yeah but that always creates a copy, even though most of the times it will not be needed. We already do this in a few places, but we need a better solution.

> ReaderWriterLockSlim is for ensure that the _listeners does not change during the copy.

`ReaderWriterLockSlim` is for managing thread safety, but this problem isn't occurring because multiple threads are accessing the collection - it's always the UI thread. The problem occurs when a classes changed listener modifies the collection that is currently being enumerated on the same thread, i.e. it's a reentrancy problem not a threading problem.